Improve the database driver selection
It detects the available database drivers for PHP on the system and
provide them as a selection list.

I removed the <optgroup> element
from the template and the array because it was a feature of the old
DBAL extension which distinguished between native SQL and ADODB support.

I added an empty option so that the user has to choose which driver
to use. After that the form reloads itself and present the credentials
login, which depends on the driver selection since Postgres needs
provides other settings like SSL.

It also saves the driver selection in LocalConfiguration.php in the
well known DB section.

Change-Id: Iedf8664283afd683bb1e7faef814804d1485345e