Make the installation of the database working again.
While the refactoring of the database connect functions I centralized
the checks for some preconditions like if PDO extension is installed
and if there is a proper database name set. With this implementation
it will throw an error is no database name is set when you call
connectDatabase from somewhere and no database name is set.

Sounds like it working as it should but no rule without an exception.

When you install TYPO3 and you like to initialize the database, TYPO3
connects to the database server to figure out if there are empty databases
which can be used. So it need to connect without a database which my implementation
prevented.

The old database solved that with calling the sql_pconnect() method directly
from the install tool and avoid the database checks. I took another approach and
asking the configuration manager if the system is currently in first installation mode
and prevent then to call the database name checks.

So I don't need to expose the method to make connections directly as API.

Change-Id: Iec1a3583e9c0abb2432f0f3f1fb1293a0f121b57