To fulfill the field type to the inserted data, change the type of boolean fields to smallint
In the absence of boolean type in MySQL, TYPO3 uses tinyint as a substitute for
booleans and also for small integer data. This leads to problems since
Doctrine maps tinyint to internal boolean type for MySQL platform.

If you create the tables from the Doctrine schema at other platforms
which support booleans as field type (e.g. Postgres), you can't insert data which is not boolean.

Change-Id: I7b466980ca97ec9d8cc5c158e4e84dc97f1dde1e